refactor: move getFieldsWithDifferences test into utils test file
