+++ app/src/modules/content/composables/use-comparison.test.ts
@@ -17,7 +17,6 @@ import api from '@/api';
 import { useComparison } from './use-comparison';
 import type { Revision } from '@/types/revisions';
 import type { ContentVersion } from '@directus/types';
-import { getFieldsWithDifferences } from '../comparison-utils';
 
 const itemVersions = {
 	data: [
@@ -213,77 +212,5 @@ describe('normalizeComparisonData', () => {
 		expect(result.incoming.things).toBe(versionComparison.data.current.things);
 		expect(result.incoming.enable).toBe(versionComparison.data.current.enable);
 		expect(result.incoming.categories).toEqual(versionComparison.data.current.categories);
-	});
-});
-
-describe('getFieldsWithDifferences', () => {
-	it('excludes related item fields when comparing revisions', () => {
-		const comparedData = {
-			outdated: false,
-			mainHash: '',
-			incoming: {
-				title: 'New Title',
-				description: 'New Description',
-				related_items: [{ id: 1, name: 'Item 1' }],
-				categories: [{ id: 2, name: 'Category 1' }],
-				status: 'published',
-			},
-			base: {
-				title: 'Old Title',
-				description: 'Old Description',
-				related_items: [{ id: 2, name: 'Item 2' }],
-				categories: [{ id: 3, name: 'Category 2' }],
-				status: 'draft',
-			},
-		};
-
-		const fieldMetadata = {
-			title: { meta: { special: [] } },
-			description: { meta: { special: [] } },
-			related_items: { meta: { special: ['m2m'] } },
-			categories: { meta: { special: ['o2m'] } },
-			status: { meta: { special: [] } },
-		};
-
-		// Test version comparison - should include all fields with differences
-		const versionFields = getFieldsWithDifferences(comparedData, fieldMetadata, 'version');
-		expect(versionFields).toContain('title');
-		expect(versionFields).toContain('description');
-		expect(versionFields).toContain('related_items');
-		expect(versionFields).toContain('categories');
-		expect(versionFields).toContain('status');
-
-		// Test revision comparison - should exclude related item fields
-		const revisionFields = getFieldsWithDifferences(comparedData, fieldMetadata, 'revision');
-		expect(revisionFields).toContain('title');
-		expect(revisionFields).toContain('description');
-		expect(revisionFields).toContain('status');
-		expect(revisionFields).not.toContain('related_items');
-		expect(revisionFields).not.toContain('categories');
-	});
-
-	it('includes related item fields when comparing versions', () => {
-		const comparedData = {
-			outdated: false,
-			mainHash: '',
-			incoming: {
-				title: 'New Title',
-				related_items: [{ id: 1, name: 'Item 1' }],
-			},
-			base: {
-				title: 'Old Title',
-				related_items: [{ id: 2, name: 'Item 2' }],
-			},
-		};
-
-		const fieldMetadata = {
-			title: { meta: { special: [] } },
-			related_items: { meta: { special: ['m2m'] } },
-		};
-
-		// Version comparison should include related item fields
-		const versionFields = getFieldsWithDifferences(comparedData, fieldMetadata, 'version');
-		expect(versionFields).toContain('title');
-		expect(versionFields).toContain('related_items');
 	});
 });
